gtk: give SC2K names to the speeds in the menu and connect them
For now, the implementation is a hack that works by changing the speed scale
on the control panel.  This will change at some point in the future when the
speed scale on the control panel is no longer needed.

The SimCity 2000 speed names are being used because they are more colorful.